assert: check prototype with loose deep equal
Currently the main reason against using the loose assertion is that
the comparison is very unreliable. The used comparison is not only
the abstract one but there are arbitrary simplifications in the
algorithms. These simplifications make it very hard to reason about
the result of the comparison.

Originally there was only a single deep equal comparison and to
overcome the shortcoming of the loose equal comparison a "strict"
version was added to the API.
To keep the original intention between the loose and the strict
version the abstract equality check is still in place for the loose
equal comparison but it is now much better to reason about the result.